+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # AWS SSO CLI Changelog
 
-## [Unreleased]
+## [v1.14.0] - 2023-10-13
+
+### Bugs
+
+ * Update net/http to v1.17.0 to fix HTTP/2 server bug CVE-2023-39325
 
 ### Changes
 
@@ -596,7 +600,8 @@
 
 Initial release
 
-[Unreleased]: https://github.com/synfinatic/aws-sso-cli/compare/v1.13.1...main
+[Unreleased]: https://github.com/synfinatic/aws-sso-cli/compare/v1.14.0...main
+[v1.14.0]: https://github.com/synfinatic/aws-sso-cli/releases/tag/v1.14.0
 [v1.13.1]: https://github.com/synfinatic/aws-sso-cli/releases/tag/v1.13.0
 [v1.13.0]: https://github.com/synfinatic/aws-sso-cli/releases/tag/v1.12.0
 [v1.12.0]: https://github.com/synfinatic/aws-sso-cli/releases/tag/v1.11.0